requirements.txt

absl-py==2.2.2
accelerate==1.6.0
attrs==25.3.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
chex==0.1.82
colorama==0.4.6
customtkinter==5.2.2
darkdetect==0.8.0
diffusers==0.32.2
docopt==0.6.2
etils==1.12.2
filelock==3.18.0
flax==0.6.9
fsspec==2025.3.2
huggingface-hub==0.30.1
idna==3.10
importlib_resources==6.5.2
jax==0.4.13
jaxlib==0.4.13
Jinja2==3.1.6
Js2Py==0.74
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
msgpack==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.24.4
opt_einsum==3.4.0
optax==0.2.1
orbax-checkpoint==0.5.16
packaging==24.2
pillow==11.1.0
pipwin==0.5.2
protobuf==6.30.2
psutil==7.0.0
PyAudio==0.2.14
pydub==0.25.1
Pygments==2.19.1
pyjsparser==2.7.1
PyPrind==2.11.3
pySmartDL==1.3.4
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
safetensors==0.5.3
scipy==1.15.2
six==1.17.0
soupsieve==2.6
SpeechRecognition==3.14.2
sympy==1.13.1
tensorstore==0.1.73
tokenizers==0.21.1
toolz==1.0.0
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.51.0
typing_extensions==4.13.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.3.0
zipp==3.21.0
